+++ __tests__/Map.ts
@@ -1,4 +1,4 @@
-import { is, List, Map, Range, Record, Seq } from 'immutable';
+import { fromJS, is, List, Map, Range, Record, Seq } from 'immutable';
 
 import * as jasmineCheck from 'jasmine-check';
 jasmineCheck.install();
@@ -85,14 +85,14 @@ describe('Map', () => {
 
   it('does not accept strings (collection, but scalar)', () => {
     expect(() => {
-      // @ts-expect-error
+      // @ts-expect-error -- constructor does not accept strings, this is expected to throw
       Map('abc');
     }).toThrow();
   });
 
   it('does not accept non-entries array', () => {
     expect(() => {
-      // @ts-expect-error
+      // @ts-expect-error -- not an array of entries, this is expected to throw
       Map([1, 2, 3]);
     }).toThrow('Expected [K, V] tuple: 1');
   });
@@ -469,6 +469,25 @@ describe('Map', () => {
     expect(m.get(a)).toBe('a');
     expect(m.get(b)).toBe('b');
     expect(m.get(c)).toBe('c');
+  });
+
+  it('supports Symbols as object constructor keys', () => {
+    const a = Symbol.for('a');
+    const b = Symbol('b');
+    const c = Symbol('c');
+    const m = Map<symbol, string>({
+      [a]: 'a',
+      [b]: 'b',
+      [c]: 'c',
+    });
+    expect(m.size).toBe(3);
+    expect(m.get(a)).toBe('a');
+    expect(m.get(b)).toBe('b');
+    expect(m.get(c)).toBe('c');
+
+    const m2 = fromJS({ [a]: 'a' }) as Map<symbol, string>;
+    expect(m2.size).toBe(1);
+    expect(m2.get(a)).toBe('a');
   });
 
   it('Symbol keys are unique', () => {